Get valueset from gateway
