+++ dgc/DCC.csproj
@@ -11,18 +11,8 @@
     <RepositoryUrl>https://github.com/ehn-digital-green-development/hcert-dotnet</RepositoryUrl>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageId>DigitalCovidCertificate</PackageId>
-    <Version>1.3.0.1</Version>
+    <Version>1.3.1.3</Version>
   </PropertyGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Valuesets\disease-agent-targeted.json" />
-    <EmbeddedResource Include="Valuesets\test-manf.json" />
-    <EmbeddedResource Include="Valuesets\test-result.json" />
-    <EmbeddedResource Include="Valuesets\test-type.json" />
-    <EmbeddedResource Include="Valuesets\vaccine-mah-manf.json" />
-    <EmbeddedResource Include="Valuesets\vaccine-medicinal-product.json" />
-    <EmbeddedResource Include="Valuesets\vaccine-prophylaxis.json" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
