--- dgc/Valuesets/disease-agent-targeted.json
@@ -1,13 +1,0 @@
-﻿{
-  "valueSetId": "disease-agent-targeted",
-  "valueSetDate": "2021-04-27",
-  "valueSetValues": {
-    "840539006": {
-      "display": "COVID-19",
-      "lang": "en",
-      "active": true,
-      "version": "http://snomed.info/sct/900000000000207008/version/20210131",
-      "system": "http://snomed.info/sct"
-    }
-  }
-}
