--- dgc/Valuesets/test-result.json
@@ -1,20 +1,0 @@
-﻿{
-  "valueSetId": "covid-19-lab-result",
-  "valueSetDate": "2021-04-27",
-  "valueSetValues": {
-    "260415000": {
-      "display": "Not detected",
-      "lang": "en",
-      "active": true,
-      "version": "http://snomed.info/sct/900000000000207008/version/20210131",
-      "system": "http://snomed.info/sct"
-    },
-    "260373001": {
-      "display": "Detected",
-      "lang": "en",
-      "active": true,
-      "version": "http://snomed.info/sct/900000000000207008/version/20210131",
-      "system": "http://snomed.info/sct"
-    }
-  }
-}
