--- dgc/Valuesets/test-type.json
@@ -1,20 +1,0 @@
-﻿{
-  "valueSetId": "covid-19-lab-test-type",
-  "valueSetDate": "2021-04-27",
-  "valueSetValues": {
-    "LP6464-4": {
-      "display": "Nucleic acid amplification with probe detection",
-      "lang": "en",
-      "active": true,
-      "version": "2.69",
-      "system": "http://loinc.org"
-    },
-    "LP217198-3": {
-      "display": "Rapid immunoassay",
-      "lang": "en",
-      "active": true,
-      "version": "2.69",
-      "system": "http://loinc.org"
-    }
-  }
-}
