--- dgc/Valuesets/vaccine-mah-manf.json
@@ -1,104 +1,0 @@
-﻿{
-  "valueSetId": "vaccines-covid-19-auth-holders",
-  "valueSetDate": "2021-04-27",
-  "valueSetValues": {
-    "ORG-100001699": {
-      "display": "AstraZeneca AB",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100030215": {
-      "display": "Biontech Manufacturing GmbH",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100001417": {
-      "display": "Janssen-Cilag International",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100031184": {
-      "display": "Moderna Biotech Spain S.L.",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100006270": {
-      "display": "Curevac AG",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100013793": {
-      "display": "CanSino Biologics",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100020693": {
-      "display": "China Sinopharm International Corp. - Beijing location",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100010771": {
-      "display": "Sinopharm Weiqida Europe Pharmaceutical s.r.o. - Prague location",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100024420": {
-      "display": "Sinopharm Zhijun (Shenzhen) Pharmaceutical Co. Ltd. - Shenzhen location",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "ORG-100032020": {
-      "display": "Novavax CZ AS",
-      "lang": "en",
-      "active": true,
-      "system": "https://spor.ema.europa.eu/v1/organisations",
-      "version": ""
-    },
-    "Gamaleya-Research-Institute": {
-      "display": "Gamaleya Research Institute",
-      "lang": "en",
-      "active": true,
-      "system": "http://ec.europa.eu/temp/vaccinemanufacturer",
-      "version": "1.0"
-    },
-    "Vector-Institute": {
-      "display": "Vector Institute",
-      "lang": "en",
-      "active": true,
-      "system": "http://ec.europa.eu/temp/vaccinemanufacturer",
-      "version": "1.0"
-    },
-    "Sinovac-Biotech": {
-      "display": "Sinovac Biotech",
-      "lang": "en",
-      "active": true,
-      "system": "http://ec.europa.eu/temp/vaccinemanufacturer",
-      "version": "1.0"
-    },
-    "Bharat-Biotech": {
-      "display": "Bharat Biotech",
-      "lang": "en",
-      "active": true,
-      "system": "http://ec.europa.eu/temp/vaccinemanufacturer",
-      "version": "1.0"
-    }
-  }
-}
